+++ baybe/recommenders/pure/bayesian/base.py
@@ -13,15 +13,14 @@ from baybe.objectives.base import Objective
 from baybe.recommenders.pure.base import PureRecommender
 from baybe.searchspace import SearchSpace
 from baybe.surrogates import CustomONNXSurrogate, GaussianProcessSurrogate
-from baybe.surrogates.base import Surrogate
-from baybe.utils.dataframe import to_tensor
+from baybe.surrogates.base import SurrogateProtocol
 
 
 @define
 class BayesianRecommender(PureRecommender, ABC):
     """An abstract class for Bayesian Recommenders."""
 
-    surrogate_model: Surrogate = field(factory=GaussianProcessSurrogate)
+    surrogate_model: SurrogateProtocol = field(factory=GaussianProcessSurrogate)
     """The used surrogate model."""
 
     acquisition_function: AcquisitionFunction = field(
@@ -51,14 +50,9 @@ class BayesianRecommender(PureRecommender, ABC):
         measurements: pd.DataFrame,
     ) -> None:
         """Create the acquisition function for the current training data."""  # noqa: E501
-        # TODO: Transition point from dataframe to tensor needs to be refactored.
-        #   Currently, surrogate models operate with tensors, while acquisition
-        #   functions with dataframes.
-        train_x = searchspace.transform(measurements, allow_extra=True)
-        train_y = objective.transform(measurements)
-        self.surrogate_model._fit(searchspace, *to_tensor(train_x, train_y))
+        self.surrogate_model.fit(searchspace, objective, measurements)
         self._botorch_acqf = self.acquisition_function.to_botorch(
-            self.surrogate_model, searchspace, train_x, train_y
+            self.surrogate_model, searchspace, objective, measurements
         )
 
     def recommend(  # noqa: D102