+++ baybe/surrogates/gaussian_process/presets/edbo.py
@@ -101,16 +101,16 @@ def _edbo_noise_factory(
 
     # low D priors
     if effective_dims < 5:
-        return [GammaPrior(1.05, 0.5), 0.1]
+        return (GammaPrior(1.05, 0.5), 0.1)
 
     # DFT optimized priors
     elif uses_descriptors and effective_dims < 100:
-        return [GammaPrior(1.5, 0.1), 5.0]
+        return (GammaPrior(1.5, 0.1), 5.0)
 
     # Mordred optimized priors
     elif uses_descriptors:
-        return [GammaPrior(1.5, 0.1), 5.0]
+        return (GammaPrior(1.5, 0.1), 5.0)
 
     # OHE optimized priors
     else:
-        return [GammaPrior(1.5, 0.1), 5.0]
+        return (GammaPrior(1.5, 0.1), 5.0)