+++ tests/test_custom_surrogate.py
@@ -6,7 +6,7 @@ import pytest
 
 from baybe import Campaign
 from baybe._optional.info import ONNX_INSTALLED
-from baybe.surrogates import CustomONNXSurrogate, register_custom_architecture
+from baybe.surrogates import CustomONNXSurrogate
 from tests.conftest import run_iterations
 
 
@@ -51,48 +51,3 @@ def test_supported_parameter_types(campaign: Campaign, should_raise: bool):
     context = pytest.raises(TypeError) if should_raise else nullcontext()
     with context:
         campaign.recommend(batch_size=1)
-
-
-def test_validate_architectures():
-    """Test architecture class validation."""
-    # Scenario: Empty Class
-    with pytest.raises(ValueError):
-        register_custom_architecture()(str)
-
-    # Scenario: Class with just `_fit`
-    with pytest.raises(ValueError):
-        register_custom_architecture()(type("PartialArch", (), {"_fit": True}))
-
-    # Scenario: Class with `_fit` and `_posterior` but not methods
-    with pytest.raises(ValueError):
-        register_custom_architecture()(
-            type("PartialArch", (), {"_fit": True, "_posterior": True})
-        )
-
-    # Scenario: Class with invalid `_fit` and `_posterior` methods
-    def _invalid_func(invalid_param1, invalid_param2=1):
-        return invalid_param1 + invalid_param2
-
-    with pytest.raises(ValueError):
-        register_custom_architecture()(
-            type(
-                "InvalidArch", (), {"_fit": _invalid_func, "_posterior": _invalid_func}
-            )
-        )
-
-    # Scenario: Class with valid `_fit` but invalid `_posterior` methods
-    def _valid_fit(self, searchspace, train_x, train_y):
-        return self and searchspace and train_x and train_y
-
-    with pytest.raises(ValueError):
-        register_custom_architecture()(
-            type("InvalidArch", (), {"_fit": _valid_fit, "_posterior": _invalid_func})
-        )
-
-    # Scenario: Both Valid
-    def _valid_posterior(self, candidates):
-        return self and candidates
-
-    register_custom_architecture()(
-        type("ValidArch", (), {"_fit": _valid_fit, "_posterior": _valid_posterior})
-    )